moved built-in function tests to builtin module
