+++ monkey-core/src/builtin.rs
@@ -25,3 +25,33 @@ pub(crate) fn get(name: &str) -> Result<BuiltinFunction, EvaluationError> {
     };
     Ok(func)
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::ast::Value;
+    use crate::environment::Environment;
+    use crate::evaluator::Evaluate;
+    use crate::lexer::Lexer;
+    use crate::parser::Parser;
+
+    #[test]
+    fn test_len_function_evaluation() {
+        let test_cases = vec![
+            ("len(\"\")", Value::Int(0)),
+            ("len(\"four\")", Value::Int(4)),
+            ("len(\"hello world\")", Value::Int(11)),
+            ("len([])", Value::Int(0)),
+            ("len([1])", Value::Int(1)),
+            ("len([1, 2])", Value::Int(2)),
+        ];
+
+        for (code, expectation) in test_cases {
+            let lexer = Lexer::new(code);
+            let mut parser = Parser::new(lexer);
+            let program = parser.parse();
+            let env = Environment::new_shared(None);
+            let return_value = program.eval(&env).unwrap();
+            assert_eq!(return_value, expectation);
+        }
+    }
+}
